Another pile of audio fixes
Various correctness and performance fixes, including a 32-bit
intermediate buffer for mixing. Still haven't hit on the root cause
of the corruption I'm seeing on hardware, but it seems related to the
mix buffer. It almost entirely disappears if the mix buffer is only
a single sample.

+++ firmware/master/common/audiomixer.h
@@ -66,7 +66,7 @@ private:
     uint32_t trackerCallbackInterval;
     uint32_t trackerCallbackCountdown;
 
-    bool mixAudio(int16_t *buffer, uint32_t numFrames);
+    bool mixAudio(int *buffer, uint32_t numFrames);
 };
 
 #endif /* AUDIOMIXER_H_ */
